Crate vsop87

This library implements the VSOP87 solutions to calculate the positions of the planets in the solar system. To use it you must include the following in your crate:

extern crate vsop87;

The main module calculates heliocentric ecliptic orbital elements for the equinox J2000.0 for the planets in the solar system, the basic VSOP87 solution. There is one module per other VSOP87 implementation: VSOP87A, VSOP87B, VSOP87C, VSOP87D and VSOP87E.

Modules

vsop87a
VSOP87A implementation

vsop87b
VSOP87B implementation

vsop87c
VSOP87C implementation

vsop87d
VSOP87D implementation

vsop87e
VSOP87E implementation

Functions

earth_moon
Calculates VSOP87 solution for Earth - Moon barycenter

jupiter
Calculates VSOP87 solution for Jupiter

keplerian_elements_from_vsop87
Calculates the keplerian orbital elements from VSOP87

mars
Calculates VSOP87 solution for Mars

mercury
Calculates VSOP87 solution for Mercury

neptune
Calculates VSOP87 solution for Neptune

saturn
Calculates VSOP87 solution for Saturn

uranus
Calculates VSOP87 solution for Uranus

venus
Calculates VSOP87 solution for Venus