Function vsop87::jupiter

pub fn jupiter(jde: f64) -> (f64, f64, f64, f64, f64, f64)

Calculates VSOP87 solution for Jupiter

This function calculates the VSOP87 solution (heliocentric ecliptic orbital elements for the equinox J2000.0) for the planet Jupiter. The parameter needed is the Julian Day Efemeris (JDE) for the given date. It returns, in order, a tuple with the values a, l, k, h, q, p of the VSOP87 solution.

Examples

let (a, l, k, h, q, p) = vsop87::jupiter(2378495.0);

assert!(a > 5.2027276672 && a < 5.2027276674);
assert!(l > 1.4820596291 && l < 1.4820596293);
assert!(k > 0.0464780412 && k < 0.0464780414);
assert!(h > 0.0116460263 && h < 0.0116460265);
assert!(q > -0.0019921307 && q < -0.0019921305);
assert!(p > 0.01123447 && p < 0.01123523);