Function vsop87::vsop87::neptune

pub fn neptune(jde: f64) -> (f64, f64, f64, f64, f64, f64)

Calculates VSOP87 solution for Neptune

This function calculates the VSOP87 solution (the heliocentric orbital elements) for the planet Neptune. The parameter needed is the Julian Day Efemeris (JDE) for the given date. It returns, in order, a tuple with the values (a, l, k, h, q, p) of the VSOP87 solution.

Examples

use vsop87::*;

let (a, l, k, h, q, p) = vsop87::neptune(2268920.0);

assert!(a > 30.1963044187 && a < 30.1963044189);
assert!(l > 5.1088676118 && l < 5.1088676120);
assert!(k > 0.0091964091 && k < 0.0091964093);
assert!(h > 0.0031103619 && h < 0.0031103621);
assert!(q > -0.0102800265 && q < -0.0102800263);
assert!(p > 0.01148076 && p < 0.01148152);