Function vsop87::vsop87::uranus

pub fn uranus(jde: f64) -> (f64, f64, f64, f64, f64, f64)

Calculates VSOP87 solution for Uranus

This function calculates the VSOP87 solution (the heliocentric orbital elements) for the planet Uranus. The parameter needed is the Julian Day Efemeris (JDE) for the given date. It returns, in order, a tuple with the values (a, l, k, h, q, p) of the VSOP87 solution.

Examples

use vsop87::*;

let (a, l, k, h, q, p) = vsop87::uranus(2232395.0);

assert!(a > 19.2497356422 && a < 19.2497356424);
assert!(l > 4.5777275752 && l < 4.5777275754);
assert!(k > -0.0466529112 && k < -0.0466529110);
assert!(h > 0.0051308956 && h < 0.0051308958);
assert!(q > 0.0019206656 && q < 0.0019206658);
assert!(p > 0.00655819 && p < 0.00655895);