Module vsop87::vsop87c

VSOP87C implementation

This module calculates heliocentric ecliptic rectangular coordinates for the equinox of the day for the planets in the solar system.

Functions

earth
Calculates VSOP87C solution for Earth

jupiter
Calculates VSOP87C solution for Jupiter

mars
Calculates VSOP87C solution for Mars

mercury
Calculates VSOP87C solution for Mercury

neptune
Calculates VSOP87C solution for Neptune

saturn
Calculates VSOP87C solution for Saturn

uranus
Calculates VSOP87C solution for Uranus

venus
Calculates VSOP87C solution for Venus