Function vsop87::vsop87c::saturn

pub fn saturn(jde: f64) -> (f64, f64, f64)

Calculates VSOP87C solution for Saturn

This function calculates the VSOP87C solution (heliocentric ecliptic rectangular coordinates for the equinox of the day) for the planet Saturn. The parameter needed is the Julian Day Efemeris (JDE) for the given date. It returns, in order, a tuple with the values x, y, z of the VSOP87C solution. Those values are the rectangular coordinates of the planet, in AU, with the Sun in the center and the ecliptic plane as reference z = 0.

Examples

use vsop87::vsop87c;

let (x, y, z) = vsop87c::saturn(2268920.0);

assert!(x > 5.9153005787 && x < 5.9153005789);
assert!(y > 6.8629464079 && y < 6.8629464081);
assert!(z > -0.34387152 && z < -0.34387092);