Module vsop87::vsop87d

VSOP87D algorithm: Heliocentric ecliptic spherical coordinates for the equinox of the day.

This module contains the functions to calculate heliocentric ecliptic spherical coordinates for the equinox of the day for the planets in the solar system.

Example

Given a date in JD, we can get the position of the planet Jupiter in the solar system using spherical coordinates. In this case, we calculate where Jupiter was in December 19th, 1399.

use vsop87::vsop87d;

let coordinates = vsop87d::jupiter(2232395.0);

assert!(coordinates.longitude() > 3.0889515349 && coordinates.longitude() < 3.0889515351);
assert!(coordinates.latitude() > 0.0231157946 && coordinates.latitude() < 0.0231157948);
assert!(coordinates.distance() > 5.44915664 && coordinates.distance() < 5.44915740);

Modules

earth

jupiter

mars

mercury

neptune

saturn

uranus

venus

Functions

earth
Calculates VSOP87D solution for Earth.

jupiter
Calculates VSOP87D solution for Jupiter.

mars
Calculates VSOP87D solution for Mars.

mercury
Calculates VSOP87D solution for Mercury.

neptune
Calculates VSOP87D solution for Neptune.

saturn
Calculates VSOP87D solution for Saturn.

uranus
Calculates VSOP87D solution for Uranus.

venus
Calculates VSOP87D solution for Venus.