Function vsop87::vsop87d::neptune

pub fn neptune(jde: f64) -> SphericalCoordinates

Calculates VSOP87D solution for Neptune.

This function calculates the VSOP87D solution (heliocentric ecliptic spherical coordinates for the equinox of the day) for the planet Neptune. The parameter needed is the Julian Day (JD) for the given date. It returns the VSOP87D solution ina SphericalCoordinates struct. Those values are the spherical coordinates of the planet, with the Sun in the center and the ecliptic plane as reference latitude = 0.

Example

Given a date in JD, we can get the position of the planet Neptune in the solar system using spherical coordinates. In this case, we calculate where Neptune was in December 19th, 1099.

use vsop87::vsop87d;

let coordinates = vsop87d::neptune(2122820.0);

assert!(coordinates.longitude() > 2.2124988266 && coordinates.longitude() < 2.2124988268);
assert!(coordinates.latitude() > 0.0027498092 && coordinates.latitude() < 0.0027498094);
assert!(coordinates.distance() > 30.06536898 && coordinates.distance() < 30.06536974);