Function vsop87::vsop87d::uranus

pub fn uranus(jde: f64) -> (f64, f64, f64)

Calculates VSOP87D solution for Uranus

This function calculates the VSOP87D solution (heliocentric ecliptic spherical coordinates for the equinox of the day) for the planet Uranus. The parameter needed is the Julian Day Efemeris (JDE) for the given date. It returns, in order, a tuple with the values l, b, r of the VSOP87D solution. Those values are the spherical coordinates of the planet, with the Sun in the center and the ecliptic plane as reference b = 0:

Ecliptic longitude (L), in radians

Ecliptic latitude (B), in radians

Radial distance (R), in AU

Examples

use vsop87::vsop87d;

let (l, b, r) = vsop87d::uranus(2159345.0);

assert!(l > 1.9333853934 && l < 1.9333853936);
assert!(b > 0.0088045917 && b < 0.0088045919);
assert!(r > 18.58414975 && r < 18.58415051);