Module vsop87::vsop87e

VSOP87E algorithm: Barycentric ecliptic rectangular coordinates for the equinox J2000.0.

This module calculates barycentric ecliptic rectangular coordinates for the equinox J2000.0 for the planets in the solar system. These coordinates are centered in the barycenter or center of masses of the solar system. This means that the Sun is not at (0, 0, 0), and its position has to be calculated independently.

Example

Given a date in JD, we can get the position of the Sun in the solar system using rectangular coordinates. In this case, we calculate where the Sun was in January 1st, 2000.

use vsop87::vsop87e;

let coordinates = vsop87e::sun(2451545.0);

assert!(coordinates.x > -0.0071415280 && coordinates.x < -0.0071415278);
assert!(coordinates.y > -0.0027881716 && coordinates.y < -0.0027881714);
assert!(coordinates.z > 0.0002041 && coordinates.z < 0.0002081);

Modules

earth

jupiter

mars

mercury

neptune

saturn

sun

uranus

venus

Functions

earth
Calculates VSOP87E solution for Earth.

jupiter
Calculates VSOP87E solution for Jupiter.

mars
Calculates VSOP87E solution for Mars.

mercury
Calculates VSOP87E solution for Mercury.

neptune
Calculates VSOP87E solution for Neptune.

saturn
Calculates VSOP87E solution for Saturn.

sun
Calculates VSOP87E solution for the Sun.

uranus
Calculates VSOP87E solution for Uranus.

venus
Calculates VSOP87E solution for Venus.