Function vsop87::vsop87e::saturn

pub fn saturn(jde: f64) -> RectangularCoordinates

Calculates VSOP87E solution for Saturn.

This function calculates the VSOP87E solution (barycentric ecliptic rectangular coordinates for the equinox J2000.0) for the planet Saturn. The parameter needed is the Julian Day (JD) for the given date. It returns the VSOP87E solution in a RectangularCoordinates structure. Those values are the rectangular coordinates of the planet, in AU, with the barycenter of the solar system in the center and the ecliptic plane as reference z = 0.

Example

Given a date in JD, we can get the position of the planet Saturn in the solar system using rectangular coordinates. In this case, we calculate where Saturn was in December 19th, 1399.

use vsop87::vsop87e;

let coordinates = vsop87e::saturn(2232395.0);

assert!(coordinates.x > 1.2645936161 && coordinates.x < 1.2645936163);
assert!(coordinates.y > -10.0240954526 && coordinates.y < -10.0240954524);
assert!(coordinates.z > 0.1345888 && coordinates.z < 0.1345928);