Module vsop87::vsop87c

VSOP87C algorithm: Heliocentric ecliptic rectangular coordinates for the equinox of the day.

This module contains the functions to calculate heliocentric ecliptic rectangular coordinates for the equinox of the day for the planets in the solar system. The most useful when converting to geocentric positions and later compute e.g. rise/set/culmination times, or the altitude and azimuth relative to your local horizon.

Example

Given a date in JD, we can get the position of the planet Venus in the solar system using rectangular coordinates. In this case, we calculate where Venus was in December 31st, 1899.

use vsop87::vsop87c;

let coordinates = vsop87c::venus(2415020.0);

assert!(coordinates.x > 0.6919778853 && coordinates.x < 0.6919778855);
assert!(coordinates.y > -0.2203045664 && coordinates.y < -0.2203045662);
assert!(coordinates.z > -0.04298775 && coordinates.z < -0.04298715);

Modules

earth

jupiter

mars

mercury

neptune

saturn

uranus

venus

Functions

earth
Calculates VSOP87C solution for Earth.

jupiter
Calculates VSOP87C solution for Jupiter.

mars
Calculates VSOP87C solution for Mars.

mercury
Calculates VSOP87C solution for Mercury.

neptune
Calculates VSOP87C solution for Neptune.

saturn
Calculates VSOP87C solution for Saturn.

uranus
Calculates VSOP87C solution for Uranus.

venus
Calculates VSOP87C solution for Venus.